+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 
+### Added
+- Support for remote script execution (via TRAMP).
+
+
 ## [1.7.1] - 2025-12-17
 ### Added
 - Option to restart REPLs with option `repl-restart` for every compile.