+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 
+## [1.7.2] - 2026-07-16
 ### Added
 - Support for remote script execution (via TRAMP).
 - Support for eshell in `flex-compile-comint`.
@@ -178,7 +179,8 @@ Major refactoring: cleaned up compilation and *package-lint* warnings.
 First major release.
 
 
-[Unreleased]: https://github.com/plandes/flex-compile/compare/v1.7.1...HEAD
+[Unreleased]: https://github.com/plandes/flex-compile/compare/v1.7.2...HEAD
+[1.7.2]: https://github.com/plandes/flex-compile/compare/v1.7.1...v1.7.2
 [1.7.1]: https://github.com/plandes/flex-compile/compare/v1.7.0...v1.7.1
 [1.7.0]: https://github.com/plandes/flex-compile/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/plandes/flex-compile/compare/v1.5...v1.6.0